requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
faiss-cpu
pypdf
wikipedia
arxiv
langchainhub